+++ packages/common/src/models/TokenStandard.ts
@@ -1,0 +1,1 @@
+export type TokenStandard = 'ERC721' | 'ERC1155'
