+++ packages/common/src/models/index.ts
@@ -27,6 +27,7 @@ export * from './Services'
 export * from './SmartCollectionVariant'
 export * from './Status'
 export * from './Stems'
+export * from './StemCategory'
 export * from './Theme'
 export * from './TimeRange'
 export * from './Timestamped'
