+++ packages/common/src/store/average-color/index.ts
@@ -1,5 +1,5 @@
 export {
   default as averageColorReducer,
-  actions as averageColorActions,
-  selectors as averageColorSelectors
+  actions as averageColorActions
 } from './slice'
+export * as averageColorSelectors from './selectors'
