+++ packages/common/src/store/average-color/slice.ts
@@ -1,11 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
-import { Nullable } from 'utils/typeUtils'
-
 import { Color } from '../../models/Color'
-import { CID } from '../../models/Identifiers'
-import { Track } from '../../models/Track'
-import { CommonState } from '../commonStore'
 
 const initialState: {
   averageColor: { [multihash: string]: Color }
@@ -43,35 +38,5 @@ const slice = createSlice({
 export const { setAverageColor, setDominantColors } = slice.actions
 
 export const actions = slice.actions
-
-export const getAverageColor = (
-  state: CommonState,
-  { multihash }: { multihash: Nullable<CID> }
-): Nullable<Color> =>
-  (multihash && state.ui.averageColor.averageColor[multihash]) || null
-
-export const getAverageColorByTrack = (
-  state: CommonState,
-  { track }: { track: Nullable<Track> }
-): Nullable<Color> => {
-  const multihash = track?.cover_art_sizes ?? track?.cover_art
-  if (!multihash) return null
-  return state.ui.averageColor.averageColor[multihash] ?? null
-}
-
-export const getDominantColorsByTrack = (
-  state: CommonState,
-  { track }: { track: Nullable<Track> }
-): Nullable<Color[]> => {
-  const multihash = track?.cover_art_sizes ?? track?.cover_art
-  if (!multihash) return null
-  return state.ui.averageColor.dominantColors[multihash] ?? null
-}
-
-export const selectors = {
-  getAverageColor,
-  getAverageColorByTrack,
-  getDominantColorsByTrack
-}
 
 export default slice.reducer
