+++ packages/common/src/store/reducers.ts
@@ -53,7 +53,7 @@ import { SettingsPageState } from './pages/settings/types'
 import smartCollection from './pages/smart-collection/slice'
 import tokenDashboardSlice from './pages/token-dashboard/slice'
 import track from './pages/track/reducer'
-import TrackPageState from './pages/track/types'
+import { TrackPageState } from './pages/track/types'
 import trendingPlaylists from './pages/trending-playlists/slice'
 import trendingUnderground from './pages/trending-underground/slice'
 import trending from './pages/trending/reducer'
