+++ packages/common/src/store/ui/modals/reducers.ts
@@ -1,14 +1,13 @@
 import { Action, combineReducers, Reducer } from '@reduxjs/toolkit'
 
 import { createChatModalReducer } from './create-chat-modal'
-import { BaseModalState } from './createModal'
 import { editPlaylistModalReducer } from './edit-playlist-modal'
 import { editTrackModalReducer } from './edit-track-modal'
 import { inboxUnavailableModalReducer } from './inbox-unavailable-modal'
 import { leavingAudiusModalReducer } from './leaving-audius-modal'
 import parentReducer, { initialState } from './parentSlice'
 import { premiumContentPurchaseModalReducer } from './premium-content-purchase-modal'
-import { Modals, ModalsState } from './types'
+import { BaseModalState, Modals, ModalsState } from './types'
 import { usdcManualTransferModalReducer } from './usdc-manual-transfer-modal'
 import { usdcPurchaseDetailsModalReducer } from './usdc-purchase-details-modal'
 import { usdcTransactionDetailsModalReducer } from './usdc-transaction-details-modal'
