+++ packages/common/src/store/ui/reactions/selectors.ts
@@ -1,0 +1,5 @@
+import { CommonState } from '../../commonStore'
+
+export const makeGetReactionForSignature =
+  (signature: string) => (state: CommonState) =>
+    state.ui.reactions.reactionsForEntityMap[signature]
